Refactored to leverage BigDecimals for precision

+++ src/main/java/com/abc/Transaction.java
@@ -2,15 +2,13 @@ package com.abc;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.util.Calendar;
-import java.util.Date;
 
 public class Transaction {
-    final double amount; // todo should use BigDecimal?
+    final BigDecimal amount;
     final String type;
     private final LocalDate transactionDate;
 
-    Transaction(final double amount, final String type) {
+    Transaction(final BigDecimal amount, final String type) {
         this.amount = amount;
         this.transactionDate = DateProvider.now();
         this.type = type;
